Pull out the guardian into a separate, reusable package
The guardian service is a component used by several contract packages.
This pulls out an operation independent version that can be used across
all of the contract families. Each family need only define the operations
specific to their contract requirements.
