+++ common-contract/MANIFEST
@@ -3,9 +3,22 @@ setup.py
 pdo/__init__.py
 pdo/contracts/common.py
 pdo/contracts/__init__.py
-pdo/contracts/jupyter/services.py
-pdo/contracts/jupyter/groups.py
-pdo/contracts/jupyter/common_widgets.py
-pdo/contracts/jupyter/utility.py
-pdo/contracts/jupyter/keys.py
+pdo/contracts/guardian/common/__init__.py
+pdo/contracts/guardian/common/capability_keys.py
+pdo/contracts/guardian/common/capability_keystore.py
+pdo/contracts/guardian/common/endpoint_registry.py
+pdo/contracts/guardian/common/guardian_service.py
+pdo/contracts/guardian/common/secrets.py
+pdo/contracts/guardian/common/utility.py
+pdo/contracts/guardian/wsgi/__init__.py
+pdo/contracts/guardian/wsgi/add_endpoint.py
+pdo/contracts/guardian/wsgi/info.py
+pdo/contracts/guardian/wsgi/process_capability.py
+pdo/contracts/guardian/wsgi/provision_token_issuer.py
+pdo/contracts/guardian/wsgi/provision_token_object.py
 pdo/contracts/jupyter/__init__.py
+pdo/contracts/jupyter/common_widgets.py
+pdo/contracts/jupyter/groups.py
+pdo/contracts/jupyter/keys.py
+pdo/contracts/jupyter/services.py
+pdo/contracts/jupyter/utility.py
