+++ common-contract/pdo/contracts/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-__all__ = [ 'common', 'jupyter' ]
+__all__ = [ 'common', 'guardian', 'jupyter' ]
