+++ common-contract/pdo/contracts/guardian/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Intel Corporation
+# Copyright 2024 Intel Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,7 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-__all__ = [
-    'ovms_predict',
-    'utility',
-    ]
+__all__ = [ 'common', 'scripts', 'wsgi']
