+++ common-contract/pdo/contracts/guardian/common/endpoint_registry.py
