+++ inference-contract/MANIFEST
@@ -1,12 +1,6 @@
 MANIFEST.in
 ./setup.py
 ./pdo/__init__.py
-./pdo/inference/wsgi/provision_token_issuer.py
-./pdo/inference/wsgi/__init__.py
-./pdo/inference/wsgi/add_endpoint.py
-./pdo/inference/wsgi/provision_token_object.py
-./pdo/inference/wsgi/info.py
-./pdo/inference/wsgi/process_capability.py
 ./pdo/inference/model_scoring_scripts/__init__.py
 ./pdo/inference/model_scoring_scripts/image_classification.py
 ./pdo/inference/model_scoring_scripts/model_scoring_script_base.py
@@ -19,16 +13,10 @@ MANIFEST.in
 ./pdo/inference/__init__.py
 ./pdo/inference/resources/resources.py
 ./pdo/inference/resources/__init__.py
-./pdo/inference/common/guardian_service.py
-./pdo/inference/common/capability_keystore.py
-./pdo/inference/common/secrets.py
 ./pdo/inference/common/__init__.py
-./pdo/inference/common/endpoint_registry.py
 ./pdo/inference/common/ovms_predict.py
 ./pdo/inference/common/utility.py
-./pdo/inference/common/capability_keys.py
 ./pdo/inference/scripts/__init__.py
-./pdo/inference/scripts/guardianCLI.py
 ./pdo/inference/scripts/scripts.py
 ./pdo/inference/jupyter.py
 ./scripts/gs_stop.sh
